+++ liver_disease.py
@@ -1,123 +1,92 @@
 import pandas as pd
 import joblib
-
+import unittest
 from sklearn.model_selection import train_test_split
 from sklearn.preprocessing import StandardScaler
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.model_selection import GridSearchCV
 from sklearn.impute import KNNImputer
-from sklearn.preprocessing import MinMaxScaler
-from sklearn.preprocessing import LabelEncoder
+from sklearn.preprocessing import MinMaxScaler, LabelEncoder
 
-location = "liver_disease_1.csv"
-data = pd.read_csv(location)
-data.shape
+def load_data(location):
+    data = pd.read_csv(location)
+    return data
 
-# Step 1 : print 2 rows for sanity check
-print(data.head(2))
+def preprocess_data(data):
+    # Step 1: Sanity check
+    print(data.head(2))
 
+    # Count the number of samples in each class
+    class_counts = data['Dataset'].value_counts()
+    print(class_counts)
 
-# Count the number of samples in each class
-class_counts = data['Dataset'].value_counts()
-print(class_counts)
+    # Label encoding
+    label_encoder = LabelEncoder()
+    data['Dataset_Encoded'] = label_encoder.fit_transform(data['Dataset'])
+    data = data.drop(['Dataset'], axis=1)
 
-data.columns
+    # Dropping columns
+    columns_to_drop = ['Total_Bilirubin', 'Alamine_Aminotransferase', 'Total_Protiens', 'Albumin_and_Globulin_Ratio']
+    data = data.drop(columns_to_drop, axis=1)
+    print('---------- After dropping columns ----------')
+    print(data.head(2))
 
-# Sample DataFrame
-df = pd.DataFrame(data)
-# Import LabelEncoder from scikit-learn
+    # Scaling data
+    scaler = MinMaxScaler()
+    data = pd.DataFrame(scaler.fit_transform(data), columns=data.columns)
 
-# Initialize the LabelEncoder
-label_encoder = LabelEncoder()
+    return data
 
-# Apply label encoding to the 'Yes_No' column
-df['Dataset_Encoded'] = label_encoder.fit_transform(df['Dataset'])
+def impute_missing_values(data):
+    df_class_0 = data.loc[data['Dataset_Encoded'] == 0]
+    df_class_1 = data.loc[data['Dataset_Encoded'] == 1]
 
-# Display the DataFrame with label encoding
-df = df.drop(['Dataset'], axis=1)
-df.shape
+    print('df_class_0.shape ', df_class_0.shape)
+    print('df_class_1.shape ', df_class_1.shape)
 
-df = df.drop(['Total_Bilirubin'], axis=1)
-df = df.drop(['Alamine_Aminotransferase'], axis=1)
-df = df.drop(['Total_Protiens'], axis=1)
-df = df.drop(['Albumin_and_Globulin_Ratio'], axis=1)
-df.columns
+    knn_imputer = KNNImputer(n_neighbors=5)
 
-# print 2 rows after column drop
-print('---------- After droppping columns ----------')
-print(df.head(2))
+    # Impute for class 0 and class 1
+    df_class_0 = pd.DataFrame(knn_imputer.fit_transform(df_class_0), columns=data.columns)
+    df_class_1 = pd.DataFrame(knn_imputer.fit_transform(df_class_1), columns=data.columns)
 
-scaler = MinMaxScaler()
-df = pd.DataFrame(scaler.fit_transform(df), columns=df.columns)
+    df_imputed = pd.concat([df_class_0, df_class_1])
 
-# Separate the data into two DataFrames based on the target variable
-df_class_0 = df.loc[df['Dataset_Encoded'] == 0]
-df_class_1 = df.loc[df['Dataset_Encoded'] == 1]
+    return df_imputed
 
-print('df_class_0.shape ', df_class_0.shape)
-print('df_class_1.shape ', df_class_1.shape)
+def split_data(df_imputed):
+    Y = df_imputed['Dataset_Encoded']
+    X = df_imputed.drop(['Dataset_Encoded'], axis=1)
+    return train_test_split(X, Y, test_size=0.2, random_state=0)
 
-# Impute missing values separately for each class using KNNImputer
-knn_imputer = KNNImputer(n_neighbors=5)
+def train_model(X_train, Y_train):
+    # Standardize features
+    scaler = StandardScaler()
+    X_train = scaler.fit_transform(X_train)
 
-# Impute for class 0
-imputer = KNNImputer(n_neighbors=5)
-df_class_0 = pd.DataFrame(
-    imputer.fit_transform(df_class_0),
-    columns=df.columns)
-df_class_1 = pd.DataFrame(
-    imputer.fit_transform(df_class_1),
-    columns=df.columns)
-df_imputed = pd.concat([df_class_0, df_class_1])
+    # Build and tune Random Forest classifier
+    rf_classifier = RandomForestClassifier(random_state=42)
+    param_grid = {'n_estimators': [5, 10, 20, 30], 'max_depth': [3, 5, 7]}
+    grid_search = GridSearchCV(estimator=rf_classifier, param_grid=param_grid, cv=5)
+    grid_search.fit(X_train, Y_train)
 
-# Split the data into features (X) and target (y)
+    print('Best hyperparameters:', grid_search.best_params_)
+    best_rf_model = grid_search.best_estimator_
 
-# Extract y vector from the dataframe
-Y = df_imputed['Dataset_Encoded']
-X = df_imputed.drop(['Dataset_Encoded'], axis=1)
+    return best_rf_model, scaler
 
+def save_model(model, scaler, model_path='best_rf_model.joblib', scaler_path='scaler.joblib'):
+    joblib.dump(model, model_path)
+    joblib.dump(scaler, scaler_path)
+    print('Model and scaler saved')
 
-# spliting train and test dataset (80:20)
-X_train, X_test, Y_train, Y_test = train_test_split(
-    X, Y, test_size=0.2, random_state=0)
+def main(location):
+    data = load_data(location)
+    data = preprocess_data(data)
+    data = impute_missing_values(data)
+    X_train, X_test, Y_train, Y_test = split_data(data)
+    model, scaler = train_model(X_train, Y_train)
+    save_model(model, scaler)
 
-# spliting train and test dataset (85:15)
-X_trainb, X_testb, Y_trainb, Y_testb = train_test_split(
-    X, Y, test_size=0.15, random_state=0)
-
-# Standardize features
-scaler = StandardScaler()
-X_train = scaler.fit_transform(X_train)
-X_test = scaler.transform(X_test)
-
-
-# Build a Random Forest classifier
-rf_classifier = RandomForestClassifier(random_state=42)
-
-# Tune hyperparameters using GridSearchCV
-# experimenting with number of trees 5,10,20,30 and maximum depth 3,5,7
-
-param_grid = {
-    'n_estimators': [5, 10, 20, 30],
-    'max_depth': [3, 5, 7],
-}
-
-grid_search = GridSearchCV(
-    estimator=rf_classifier,
-    param_grid=param_grid,
-    cv=5)
-grid_search.fit(X_train, Y_train)
-
-# Print the best hyperparameters found from above grid search.
-print('Best hyperparameters:',  grid_search.best_params_)
-
-best_rf_model = grid_search.best_estimator_
-
-
-print('Training Liver disease model completed')
-print('saving model')
-
-joblib.dump(best_rf_model, 'best_rf_model.joblib')
-joblib.dump(scaler, 'scaler.joblib')
-
-print('Model saved')
+if __name__ == '__main__':
+    main("liver_disease_1.csv")
